Use Antd "Notification" rather than "Alert" (member page, privacy switch):
The "Alert" component was breaking the layout. Andt notification is global and does not affect the layout.
Secondly, show notification every time an error is thrown when trying to toggle (before it was only shown once)
Eventually, the notification will have to be extracted from the component and be managed globally by the app.